requirements.txt
PyQt6
requests
pyinstaller
Pillow
dotenv
vdf